Update README.md
# The Meal App 
This project is a web application that utilizes the power of vanilla JavaScript to provide a seamless user experience. The main feature of the app is its ability to search for food recipes using the Themealdb API. The API returns a variety of data related to the searched recipe, such as ingredients, instructions, and nutritional information. The app then displays this information in an easy-to-read format, allowing users to quickly find the recipe they are looking for.

In addition to the search feature, the app also includes the ability for users to view details of the selected recipe. This includes a more in-depth look at the recipe's ingredients and instructions, as well as nutritional information. Users can also mark their favorite recipes for easy access in the future.

Overall, this project demonstrates your skills in working with APIs, JavaScript, and web development in general. The use of vanilla JavaScript allows for a lightweight and fast-performing application, without the need for any additional JavaScript frameworks. It's a great addition to your portfolio to show your ability to create a functional web application from scratch.


## Key Features

- Search Meal by name
- Auto sugestion on pressed keyword
- Save favrouite meals for future use
- Ability to restore saved meals even after reopening the browser
- Works seamlessly on both mobile and desktop devices

## Getting Started

These instructions will get you a copy of the project up and running on your local machine for development and testing purposes.

### Prerequisites

- A web browser (e.g. Google Chrome, Firefox, Safari)
- An internet connection
- A code editor (e.g. VS Code, Atom)

### Installing

1. Clone or download the repository to your local machine
2. Open the project in your code editor
3. Open the index.html file in your web browser to run the app

## Built With

- HTML
- CSS
- JavaScript